requirements.txt
ormdantic
fastapi
uvicorn
pydantic
python-decouple
starlette
aiosqlite
authx==0.7.0
